Use Aria2 downloader to speed large file downloads and allow continue if transfer interrupted
Decided there were still issues with the Invoke-FileDownload function as the file could be done downloading and the progressbar would keep going. Not good for speed. So opted to spend 2Mb grabbing the Aria2 downloader software and use it to download everything after that. Seems to work well with powerful full speed download and a nice progress system.